Stop wrapping end-of-line comments.
This change treats end-of-line comments (i.e., those that are preceded by
other tokens on the same line) as if they have a size of zero, causing them
to not participate in wrapping. Finding the right place to place these is
challenging, so instead, we emit diagnostics instead that tell the user they
should move them.