Modify the pretty printer to use an explicit break-based model.
Instead of using consistent/inconsistent groups to determine when a line
should wrap, we now specify the behavior on the breaks and introduce a
few break types that more closely align with the regular
line/continuation line behavior that we want.

I also took the opportunity here to get rid of the old debug mode,
because the --token-stream mode introduced by dabelknap@ is far better.