Enable the pretty-printer pass by default.
It can be disabled (intended for development/debugging purposes only)
using the `--debug-disable-pretty-print` option.

This change also renames `--print-token-stream` to
`--debug-dump-token-stream`; going forward, we should namespace all
options intended for debugging with the prefix `--debug-*`.

These debugging options intentionally have no "usage" strings, which
hides them from the `--help` screen. In a future PR, we can update the
README to mention them (it needs a bigger overhaul that I don't want to
include in this PR).