Allow breaks to ignore discretionary newlines.
Previously, discretionary newlines could appear anywhere that a break was
allowed. Some breaks, however, exist only as "last resorts" to allow a
line break to occur when there is no other way to break the tokens before
it. To handle this, we add a flag to breaks that allows them to discard
the user's newline unless that last resort behavior occurs.